add alert on delete todo

+++ lib/features/todo_list/todo_list_screen.dart
@@ -8,6 +8,7 @@ import 'package:todo_clean_solid/features/todo_list/cubit/todo_cubit.dart';
 import 'package:todo_clean_solid/models/todo.dart';
 import 'package:provider/provider.dart';
 import 'package:todo_clean_solid/extension/quicksand_text_style.dart';
+import 'package:todo_clean_solid/widgets/custom_dialog.dart';
 import 'package:todo_clean_solid/widgets/priority_dropdown.dart';
 import 'package:uuid/uuid.dart';
 
@@ -85,7 +86,20 @@ class TodoListScreen extends StatelessWidget {
                                 trailing: IconButton(
                                   icon: const Icon(Icons.close),
                                   onPressed: () {
-                                    context.read<TodoCubit>().deleteTodo(i);
+                                    showReusableAlertDialog(
+                                      context: context,
+                                      title: 'Delete Todo',
+                                      content: 'Are you sure?',
+                                      confirmButtonText: 'YES',
+                                      cancelButtonText: 'NO',
+                                      onConfirm: () {
+                                        context.read<TodoCubit>().deleteTodo(i);
+                                        context.pop();
+                                      },
+                                      onCancel: () {
+                                        context.pop();
+                                      },
+                                    );
                                   },
                                 ),
                               ),
